[nano.wang][Roam]:Enhance the EAPOL handshake offload INI per AKM and Add the support for SAE EAPOL handshake offload
<Change Type>: Inner requirements
<Root Cause>:
Existing INI disable 4way_hs_offload is defined as a boolean
that contorls the EAPOL Handshake to be offloaded to Firmware or
Not. There is no way to specify the EAPOL offload per AKM.

Hence extended the support of EAPOL handshake offload to firmware
per AKM by convertig this INI to a bitmap integer. With INI value
as 0 by default Firmware takes care of 4-way HS during roam for all
AKMs except for SAE and OWE. If this INI value is 1 then Host takes
care of EAPOL for all AKMs.

Having the backward compatibility, added a new value to INI. If
this INI value is set to 2 and also if the SAE EAPOL offload service
bit is set by Firmware then the behavior is same as INI with 0 except
that Firmware takes care of 4-way HS for SAE roam as well. And for
OWE it is still handled by supplicant.

To achieve this, new roam offload flag is introduced to indicate
Firmware about SAE roam EAPOL offload, so that in roam sync
indication Firmware will inform the status as connected but not
authenticated so that supplicant can take care of 4-way EAPOL HS.

Set the SAE EAPOL offload support in resource config based on
WMI_SERVICE_SAE_EAPOL_OFFLOAD_SUPPORT bit received in service
bitmap from Firmware.
<Jira ID>: NA
<RN Content>:N

<<< Test Notes >>>
<Test-Proposal>:N
<Stress-Test>:N

Change-Id: Ia542ce143c745149455066b8faae7bb39650435e